Add alwaysSetCookie option
Causes a Set-Cookie header to be sent on every response, regardless of if the
session has been accessed, populated, or changed. This results in the
expiration date being updated on every request.

Similar to express-session's rolling option. See
https://github.com/expressjs/session#rolling

Also see https://github.com/senchalabs/connect/issues/670 for more context and
use cases for this feature.