Enable setting multiple cookie-sessions
Introduces a new option, 'sessionName', which is the key under
`Express.Request.sessions` where the session can be accessed. 'session' is the
default sessionName, and sessions with sessionName 'session' are still
accessible at `Express.Request.session`. This API was chosen to avoid breaking
existing applications.

Rejected alternative APIs:
- The `name` property acts as the session accessor on `req.sessions`. Either
obviating the need for the `sessionName` option, or perhaps allowing the
`sessionName` to default to `name`. This would be a breaking change in all
apps that provide a non-default value for `name`. I considered working around
this by always making the first cookie-session created the one accessible at
`req.session`, but this felt hacky and potentially confusing.
- Each session is accessed by adding its sessionName as a property on `req`.
I feared this would pollute the namespace, risking name collisions with other
modules.